Crate brillig

The Brillig bytecode is distinct from regular ACIR in that it does not generate constraints.

Modules

black_box 🔒

foreign_call 🔒

opcodes 🔒

Structs

ForeignCallResult

Represents the full output of a foreign call.

HeapArray

A fixed-sized array starting from a Brillig memory location.

HeapVector

A memory-sized vector passed starting from a Brillig memory location and with a memory-held size

Enums

BinaryFieldOp

Binary fixed-length field expressions

BinaryIntOp

Binary fixed-length integer expressions

BitSize

BlackBoxOp

These opcodes provide an equivalent of ACIR blackbox functions. They are implemented as native functions in the VM.

ForeignCallParam

Single output of a foreign call.

HeapValueType

Describes the memory layout for an array/vector element

IntegerBitSize

MemoryAddress

Represents an address in the VM’s memory. Supports both direct and relative addressing.

Opcode

ValueOrArray

Lays out various ways an external foreign call’s input and output data may be interpreted inside Brillig. This data can either be an individual value or memory.

Type Aliases

Label