brillig
Enum BlackBoxOp

pub enum BlackBoxOp {
    AES128Encrypt {
        inputs: HeapArray,
        iv: HeapArray,
        key: HeapArray,
        outputs: HeapArray,
    },
    Blake2s {
        message: HeapArray,
        output: HeapArray,
    },
    Blake3 {
        message: HeapArray,
        output: HeapArray,
    },
    Keccakf1600 {
        input: HeapArray,
        output: HeapArray,
    },
    EcdsaSecp256k1 {
        hashed_msg: HeapArray,
        public_key_x: HeapArray,
        public_key_y: HeapArray,
        signature: HeapArray,
        result: MemoryAddress,
    },
    EcdsaSecp256r1 {
        hashed_msg: HeapArray,
        public_key_x: HeapArray,
        public_key_y: HeapArray,
        signature: HeapArray,
        result: MemoryAddress,
    },
    MultiScalarMul {
        points: HeapArray,
        scalars: HeapArray,
        outputs: HeapArray,
    },
    EmbeddedCurveAdd {
        input1_x: MemoryAddress,
        input1_y: MemoryAddress,
        input1_infinite: MemoryAddress,
        input2_x: MemoryAddress,
        input2_y: MemoryAddress,
        input2_infinite: MemoryAddress,
        result: HeapArray,
    },
    Poseidon2Permutation {
        message: HeapArray,
        output: HeapArray,
    },
    Sha256Compression {
        input: HeapArray,
        hash_values: HeapArray,
        output: HeapArray,
    },
    ToRadix {
        input: MemoryAddress,
        radix: MemoryAddress,
        output_pointer: MemoryAddress,
        num_limbs: MemoryAddress,
        output_bits: MemoryAddress,
    },
}

These opcodes provide an equivalent of ACIR blackbox functions. They are implemented as native functions in the VM. For more information, see the ACIR blackbox functions in acir::circuit::opcodes::BlackBoxFuncCall

Variants

AES128Encrypt

Encrypts a message using AES-128.

Fields

inputs: HeapArray

iv: HeapArray

key: HeapArray

outputs: HeapArray

Blake2s

Calculates the Blake2s hash of the inputs.

Fields

message: HeapArray

output: HeapArray

Blake3

Calculates the Blake3 hash of the inputs.

Fields

message: HeapArray

output: HeapArray

Keccakf1600

Keccak permutation function of 1600 width.

Fields

input: HeapArray

output: HeapArray

EcdsaSecp256k1

Verifies an ECDSA signature over the secp256k1 curve.

Fields

hashed_msg: HeapArray

public_key_x: HeapArray

public_key_y: HeapArray

signature: HeapArray

result: MemoryAddress

EcdsaSecp256r1

Verifies an ECDSA signature over the secp256r1 curve.

Fields

hashed_msg: HeapArray

public_key_x: HeapArray

public_key_y: HeapArray

signature: HeapArray

result: MemoryAddress

MultiScalarMul

Performs multi scalar multiplication over the embedded curve.

Fields

points: HeapArray

scalars: HeapArray

outputs: HeapArray

EmbeddedCurveAdd

Performs addition over the embedded curve.

Fields

input1_x: MemoryAddress

input1_y: MemoryAddress

input1_infinite: MemoryAddress

input2_x: MemoryAddress

input2_y: MemoryAddress

input2_infinite: MemoryAddress

result: HeapArray

Poseidon2Permutation

Applies the Poseidon2 permutation function to the given state, outputting the permuted state.

Fields

message: HeapArray

output: HeapArray

Sha256Compression

Applies the SHA-256 compression function to the input message

Fields

input: HeapArray

hash_values: HeapArray

output: HeapArray

ToRadix

Returns a decomposition in num_limbs limbs of the given input over the given radix.

The value stored in radix must be in the range [2, 256]

num_limbs must be at least one if the value stored in input is not zero.

The value stored in output_bits must have a bit_size of one. That value specifies whether we should decompose into bits. The value stored in the radix address must be two if the value stored in output_bits is equal to one.

Native to the Brillig VM and not supported as an ACIR black box function.

Fields

input: MemoryAddress

radix: MemoryAddress

output_pointer: MemoryAddress

num_limbs: MemoryAddress

output_bits: MemoryAddress