brillig
Module opcodes

Structs

HeapArray

A fixed-sized array starting from a Brillig memory location.

HeapVector

A memory-sized vector passed starting from a Brillig memory location and with a memory-held size.

Enums

BinaryFieldOp

Binary operations on field elements.

BinaryIntOp

Binary fixed-length integer expressions

BitSize

Represents the bit size of values in Brillig.

BrilligOpcode

HeapValueType

Describes the memory layout for an array/vector element

IntegerBitSize

Represents the bit size of unsigned integer types in Brillig.

MemoryAddress

Represents an address in the VM’s memory. Supports both direct and relative addressing.

ValueOrArray

Lays out various ways an external foreign call’s input and output data may be interpreted inside Brillig. This data can either be an individual value or memory.

Type Aliases

Label

Represents a program location (instruction index) used as a jump target.